+++ module/Motion/InCrossRight.h
@@ -38,7 +38,7 @@ class InCrossRight : public BlockAreaMotion {
   double targetDistance;    // 目標距離 (25.0mm)
   double dsTargetSpeed;     // 距離指定直進の目標速度 [mm/s]
   int targetAngle;          // 目標回頭角度 (74°)
-  double prPwm;             // 角度指定回頭の目標PWM
+  int prPwm;                // 角度指定回頭の目標PWM
   bool isClockwise = true;  // 回頭の方向は右回り
   bool isLeftEdge = true;   // 回頭前のエッジは左
   bool nextEdge = false;    // 回頭後のエッジを右に変更
